updates in whatsapp bot

+++ run_docker.sh
@@ -1,43 +1,57 @@
 #!/bin/bash
 
-# Argument to be passed to the Python script inside the Docker container
+if [ -f .env ]; then
+    set -o allexport
+    source .env
+    set +o allexport
+else
+    echo ".env file not found!"
+    exit 1
+fi
+
+if [ "$ENABLE_AGENTIC_RAG" == "True" ]; then
+    CONTAINER_NAME="agentic-rag-bot"
+else
+    CONTAINER_NAME="ragtime"
+fi
+
 BOT_NAME=$1
 
-# Network name
 NETWORK_NAME=ragtime-net
 
-# Create Docker network if it doesn't exist
 docker network inspect $NETWORK_NAME &>/dev/null || docker network create $NETWORK_NAME
 
-# Build docker container
-docker build -f Dockerfile . --tag=ragtime-bot:latest
+docker build -f Dockerfile . --tag=${CONTAINER_NAME}:latest
 
-# Remove old container if it exists
-docker container inspect ragtime-bot &>/dev/null && docker rm -f ragtime-bot
+docker container inspect ${CONTAINER_NAME} &>/dev/null && docker rm -f ${CONTAINER_NAME}
 
-# Run the ragtime-bot container on the same network
-if [ -z "$RUN_BOT_ARG" ]; then
+if [ -z "$BOT_NAME" ]; then
   docker run -d \
     --network=$NETWORK_NAME \
     -v "$(pwd)/data/convo.db:/app/data/convo.db" \
+    -p 5000:5000 \
     --env-file ./.env \
-    --name ragtime-bot \
-    ragtime-bot
+    --name ${CONTAINER_NAME} \
+    ${CONTAINER_NAME}
 else
   docker run -d \
     --network=$NETWORK_NAME \
     -v "$(pwd)/data/convo.db:/app/data/convo.db" \
+    -p 5000:5000 \
     --env-file ./.env \
-    --name ragtime-bot \
-    ragtime-bot python main.py "$BOT_NAME"
+    --name ${CONTAINER_NAME} \
+    ${CONTAINER_NAME} python main.py "$BOT_NAME"
 fi
 
-# Remove old redis-server container if it exists
-docker container inspect redis-server &>/dev/null && docker rm -f redis-server
+if [ "$ENABLE_AGENTIC_RAG" == "True" ]; then
 
-# Run redis-server on the same network
-docker run -d \
-  --network=$NETWORK_NAME \
-  -p 6379:6379 \
-  --name redis-server \
-  redis
+    docker container inspect redis-server &>/dev/null && docker rm -f redis-server
+
+    docker run -d \
+      --network=$NETWORK_NAME \
+      -p 6379:6379 \
+      --name redis-server \
+      redis
+else
+    echo "Skipping redis-server container as ENABLE_AGENTIC_RAG is not set to true."
+fi
